add omp ver
