+++ tasks/omp/gnitienko_k_strassen_alg_omp/src/ops_omp.cpp
@@ -1,11 +1,13 @@
-#include "seq/gnitienko_k_strassen_alg/include/ops_seq.hpp"
+#include "omp/gnitienko_k_strassen_alg_omp/include/ops_omp.hpp"
+
+#include <omp.h>
 
 #include <cmath>
 #include <cstddef>
-#include <utility>
+#include <iostream>
 #include <vector>
 
-bool gnitienko_k_strassen_algorithm::StrassenAlgSeq::PreProcessingImpl() {
+bool gnitienko_k_strassen_algorithm_omp::StrassenAlgOpenMP::PreProcessingImpl() {
   size_t input_size = task_data->inputs_count[0];
   auto* in_ptr = reinterpret_cast<double*>(task_data->inputs[0]);
   input_1_ = std::vector<double>(in_ptr, in_ptr + input_size);
@@ -40,29 +42,30 @@ bool gnitienko_k_strassen_algorithm::StrassenAlgSeq::PreProcessingImpl() {
   return true;
 }
 
-bool gnitienko_k_strassen_algorithm::StrassenAlgSeq::ValidationImpl() {
+bool gnitienko_k_strassen_algorithm_omp::StrassenAlgOpenMP::ValidationImpl() {
   return task_data->inputs_count[0] == task_data->outputs_count[0];
 }
 
-void gnitienko_k_strassen_algorithm::StrassenAlgSeq::AddMatrix(const std::vector<double>& a,
-                                                               const std::vector<double>& b, std::vector<double>& c,
-                                                               int size) {
+void gnitienko_k_strassen_algorithm_omp::StrassenAlgOpenMP::AddMatrix(const std::vector<double>& a,
+                                                                      const std::vector<double>& b,
+                                                                      std::vector<double>& c, int size) {
   for (int i = 0; i < size * size; ++i) {
     c[i] = a[i] + b[i];
   }
 }
 
-void gnitienko_k_strassen_algorithm::StrassenAlgSeq::SubMatrix(const std::vector<double>& a,
-                                                               const std::vector<double>& b, std::vector<double>& c,
-                                                               int size) {
+void gnitienko_k_strassen_algorithm_omp::StrassenAlgOpenMP::SubMatrix(const std::vector<double>& a,
+                                                                      const std::vector<double>& b,
+                                                                      std::vector<double>& c, int size) {
   for (int i = 0; i < size * size; ++i) {
     c[i] = a[i] - b[i];
   }
 }
 
-void gnitienko_k_strassen_algorithm::StrassenAlgSeq::TrivialMultiply(const std::vector<double>& a,
-                                                                     const std::vector<double>& b,
-                                                                     std::vector<double>& c, int size) {
+void gnitienko_k_strassen_algorithm_omp::StrassenAlgOpenMP::TrivialMultiply(const std::vector<double>& a,
+                                                                            const std::vector<double>& b,
+                                                                            std::vector<double>& c, int size) {
+#pragma omp parallel for schedule(static)
   for (int i = 0; i < size; ++i) {
     for (int j = 0; j < size; ++j) {
       c[(i * size) + j] = 0;
@@ -73,9 +76,9 @@ void gnitienko_k_strassen_algorithm::StrassenAlgSeq::TrivialMultiply(const std::
   }
 }
 
-void gnitienko_k_strassen_algorithm::StrassenAlgSeq::StrassenMultiply(const std::vector<double>& a,
-                                                                      const std::vector<double>& b,
-                                                                      std::vector<double>& c, int size) {
+void gnitienko_k_strassen_algorithm_omp::StrassenAlgOpenMP::StrassenMultiply(const std::vector<double>& a,
+                                                                             const std::vector<double>& b,
+                                                                             std::vector<double>& c, int size) {
   if (size <= TRIVIAL_MULTIPLICATION_BOUND_) {
     TrivialMultiply(a, b, c, size);
     return;
@@ -92,6 +95,7 @@ void gnitienko_k_strassen_algorithm::StrassenAlgSeq::StrassenMultiply(const std:
   std::vector<double> b21(half_size * half_size);
   std::vector<double> b22(half_size * half_size);
 
+#pragma omp parallel for schedule(static)
   for (int i = 0; i < half_size; ++i) {
     for (int j = 0; j < half_size; ++j) {
       a11[(i * half_size) + j] = a[(i * size) + j];
@@ -101,6 +105,7 @@ void gnitienko_k_strassen_algorithm::StrassenAlgSeq::StrassenMultiply(const std:
     }
   }
 
+#pragma omp parallel for schedule(static)
   for (int i = 0; i < half_size; ++i) {
     for (int j = 0; j < half_size; ++j) {
       b11[(i * half_size) + j] = b[(i * size) + j];
@@ -118,39 +123,77 @@ void gnitienko_k_strassen_algorithm::StrassenAlgSeq::StrassenMultiply(const std:
   std::vector<double> v1(half_size * half_size);
   std::vector<double> v2(half_size * half_size);
 
-  // d = (a11 + a22) * (b11 + b22)
-  std::vector<double> temp_a(half_size * half_size);
-  std::vector<double> temp_b(half_size * half_size);
+#pragma omp parallel num_threads(7)
+  {
+#pragma omp sections
+    {
+#pragma omp section
+      {
+        // d = (a11 + a22) * (b11 + b22)
+        std::vector<double> temp_a(half_size * half_size);
+        std::vector<double> temp_b(half_size * half_size);
+        AddMatrix(a11, a22, temp_a, half_size);
+        AddMatrix(b11, b22, temp_b, half_size);
+        StrassenMultiply(temp_a, temp_b, d, half_size);
+      }
 
-  AddMatrix(a11, a22, temp_a, half_size);
-  AddMatrix(b11, b22, temp_b, half_size);
-  StrassenMultiply(temp_a, temp_b, d, half_size);
+#pragma omp section
+      {
+        std::vector<double> temp_a(half_size * half_size);
+        std::vector<double> temp_b(half_size * half_size);
+        // d1 = (a12 - a22) * (b21 + b22)
+        SubMatrix(a12, a22, temp_a, half_size);
+        AddMatrix(b21, b22, temp_b, half_size);
+        StrassenMultiply(temp_a, temp_b, d1, half_size);
+      }
 
-  // d1 = (a12 - a22) * (b21 + b22)
-  SubMatrix(a12, a22, temp_a, half_size);
-  AddMatrix(b21, b22, temp_b, half_size);
-  StrassenMultiply(temp_a, temp_b, d1, half_size);
+#pragma omp section
+      {
+        std::vector<double> temp_a(half_size * half_size);
+        std::vector<double> temp_b(half_size * half_size);
+        // d2 = (a21 - a11) * (b11 + b12)
+        SubMatrix(a21, a11, temp_a, half_size);
+        AddMatrix(b11, b12, temp_b, half_size);
+        StrassenMultiply(temp_a, temp_b, d2, half_size);
+      }
 
-  // d2 = (a21 - a11) * (b11 + b12)
-  SubMatrix(a21, a11, temp_a, half_size);
-  AddMatrix(b11, b12, temp_b, half_size);
-  StrassenMultiply(temp_a, temp_b, d2, half_size);
+#pragma omp section
+      {
+        std::vector<double> temp_a(half_size * half_size);
+        std::vector<double> temp_b(half_size * half_size);
+        // h1 = (a11 + a12) * b22
+        AddMatrix(a11, a12, temp_a, half_size);
+        StrassenMultiply(temp_a, b22, h1, half_size);
+      }
 
-  // h1 = (a11 + a12) * b22
-  AddMatrix(a11, a12, temp_a, half_size);
-  StrassenMultiply(temp_a, b22, h1, half_size);
+#pragma omp section
+      {
+        std::vector<double> temp_a(half_size * half_size);
+        std::vector<double> temp_b(half_size * half_size);
+        // h2 = (a21 + a22) * b11
+        AddMatrix(a21, a22, temp_a, half_size);
+        StrassenMultiply(temp_a, b11, h2, half_size);
+      }
 
-  // h2 = (a21 + a22) * b11
-  AddMatrix(a21, a22, temp_a, half_size);
-  StrassenMultiply(temp_a, b11, h2, half_size);
+#pragma omp section
+      {
+        std::vector<double> temp_a(half_size * half_size);
+        std::vector<double> temp_b(half_size * half_size);
+        // v1 = a22 * (b21 - b11)
+        SubMatrix(b21, b11, temp_b, half_size);
+        StrassenMultiply(a22, temp_b, v1, half_size);
+      }
 
-  // v1 = a22 * (b21 - b11)
-  SubMatrix(b21, b11, temp_b, half_size);
-  StrassenMultiply(a22, temp_b, v1, half_size);
-
-  // v2 = a11 * (b12 - b22)
-  SubMatrix(b12, b22, temp_b, half_size);
-  StrassenMultiply(a11, temp_b, v2, half_size);
+#pragma omp section
+      {
+        std::vector<double> temp_a(half_size * half_size);
+        std::vector<double> temp_b(half_size * half_size);
+        // v2 = a11 * (b12 - b22)
+        SubMatrix(b12, b22, temp_b, half_size);
+        StrassenMultiply(a11, temp_b, v2, half_size);
+      }
+    }
+  }
 
   std::vector<double> c11(half_size * half_size);
   std::vector<double> c12(half_size * half_size);
@@ -166,6 +209,7 @@ void gnitienko_k_strassen_algorithm::StrassenAlgSeq::StrassenMultiply(const std:
   AddMatrix(c22, v2, c22, half_size);
   SubMatrix(c22, h2, c22, half_size);
 
+#pragma omp parallel for schedule(static)
   for (int i = 0; i < half_size; ++i) {
     for (int j = 0; j < half_size; ++j) {
       c[(i * size) + j] = c11[(i * half_size) + j];
@@ -176,7 +220,7 @@ void gnitienko_k_strassen_algorithm::StrassenAlgSeq::StrassenMultiply(const std:
   }
 }
 
-bool gnitienko_k_strassen_algorithm::StrassenAlgSeq::RunImpl() {
+bool gnitienko_k_strassen_algorithm_omp::StrassenAlgOpenMP::RunImpl() {
   StrassenMultiply(input_1_, input_2_, output_, size_);
   if (extend_ != 0) {
     int original_size = size_ - extend_;
@@ -193,9 +237,10 @@ bool gnitienko_k_strassen_algorithm::StrassenAlgSeq::RunImpl() {
   return true;
 }
 
-bool gnitienko_k_strassen_algorithm::StrassenAlgSeq::PostProcessingImpl() {
+bool gnitienko_k_strassen_algorithm_omp::StrassenAlgOpenMP::PostProcessingImpl() {
   for (size_t i = 0; i < output_.size(); i++) {
     reinterpret_cast<double*>(task_data->outputs[0])[i] = round(output_[i] * 10000) / 10000;
+    ;
   }
   return true;
 }
