--- tasks/seq/gnitienko_k_strassen_alg/data/test.txt
@@ -1,1 +1,0 @@
-100
